media: i2c: crlmodule: fix compile warnings

+++ drivers/media/i2c/crlmodule/crlmodule-nvm.c
@@ -26,7 +26,7 @@ static ssize_t crlmodule_sysfs_nvm_read(struct device *dev,
 	struct v4l2_subdev *subdev = i2c_get_clientdata(to_i2c_client(dev));
 	struct crl_sensor *sensor = to_crlmodule_sensor(subdev);
 
-	memcpy(buf, sensor->nvm_data, min(PAGE_SIZE, sensor->nvm_size));
+	memcpy(buf, sensor->nvm_data, min(((int)PAGE_SIZE), ((int)sensor->nvm_size)));
 	return sensor->nvm_size;
 }
